+++ src/main/java/nl/tudelft/contextproject/Level/MazeTile.java
@@ -1,4 +1,4 @@
-package nl.tudelft.contextproject;
+package nl.tudelft.contextproject.Level;
 
 import com.jme3.asset.AssetManager;
 import com.jme3.material.Material;
